requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
youtube-transcript-api==1.2.3
yt-dlp==2023.12.30
python-dotenv==1.0.0
firebase-admin==6.3.0
httpx==0.27.0
google-api-python-client
isodate